+++ interfaces/carInterface.ts
@@ -1,0 +1,8 @@
+export default interface CarModel {
+    id: string;
+    name: string;
+    image: string;
+    specs?: {
+        dailyPrice?: number;
+    };
+}
