+++ package.json
@@ -20,28 +20,28 @@
     "@react-navigation/native": "^7.0.14",
     "@shopify/restyle": "^2.4.5",
     "@tanstack/react-query": "^5.76.1",
-    "expo": "^53.0.9",
-    "expo-blur": "~14.1.4",
+    "expo": "53.0.10",
+    "expo-blur": "~14.1.5",
     "expo-constants": "~17.1.6",
     "expo-font": "~13.3.1",
     "expo-haptics": "~14.1.4",
-    "expo-linear-gradient": "~14.1.4",
+    "expo-linear-gradient": "~14.1.5",
     "expo-linking": "~7.1.5",
     "expo-router": "~5.0.7",
     "expo-secure-store": "~14.2.3",
-    "expo-splash-screen": "~0.30.8",
+    "expo-splash-screen": "~0.30.9",
     "expo-status-bar": "~2.2.3",
-    "expo-symbols": "~0.4.4",
-    "expo-system-ui": "~5.0.7",
-    "expo-updates": "~0.28.13",
+    "expo-symbols": "~0.4.5",
+    "expo-system-ui": "~5.0.8",
+    "expo-updates": "~0.28.14",
     "expo-web-browser": "~14.1.6",
     "react": "19.0.0",
     "react-dom": "19.0.0",
-    "react-native": "0.79.2",
+    "react-native": "0.79.3",
     "react-native-gesture-handler": "~2.24.0",
     "react-native-reanimated": "~3.17.4",
     "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.10.0",
+    "react-native-screens": "~4.11.1",
     "react-native-web": "^0.20.0",
     "react-native-webview": "13.13.5"
   },
@@ -51,7 +51,7 @@
     "@types/react": "~19.0.10",
     "@types/react-test-renderer": "^18.3.0",
     "jest": "^29.2.1",
-    "jest-expo": "~53.0.5",
+    "jest-expo": "~53.0.6",
     "typescript": "^5.3.3"
   },
   "private": true
